+++ roomie_board/src/components/Calendar 2.css
@@ -1,0 +1,60 @@
+.fc-timegrid-slot-lane,
+.fc-timegrid-slot,
+.fc-timegrid-col,
+.fc-timegrid-axis,
+.fc-timegrid-col-frame {
+  border: none !important;
+}
+.fc-event {
+    border-radius: 12px;
+    background-color: #eef3ff;
+    border: 1px solid #d6e0f5;
+    padding: 8px;
+    color: #000;
+    font-size: 12px;
+    display: flex;
+    flex-direction: column;
+    gap: 4px;
+    cursor: pointer;
+}
+
+.fc-view-harness {
+    background-color: #ffffff;
+}
+
+.fc-event-title {
+    font-weight: bold;
+    color: #5a67d8;
+    margin: 0;
+}
+  
+.fc-event-time {
+    font-size: 10px;
+    color: #718096;
+    margin: 0;
+}
+
+.event-details {
+    display: flex;
+    flex-direction: column;
+    gap: 4px;
+  }
+  
+.event-details span {
+    font-size: 10px;
+    color: #4a5568;
+}
+  
+.assignee {
+    display: flex;
+    align-items: center;
+    gap: 4px;
+}
+
+.assignee span {
+    background-color: #dceafe;
+    border-radius: 50%;
+    padding: 4px;
+    font-weight: bold;
+    color: #3182ce;
+}
